Move DatetimePrecision to aiida.common
