+++ aiida/restapi/common/utils.py
@@ -15,6 +15,7 @@ from flask import jsonify
 from flask.json import JSONEncoder
 from wrapt import decorator
 
+from aiida.common.utils import DatetimePrecision
 from aiida.common.exceptions import InputValidationError, ValidationError
 from aiida.manage.manager import get_manager
 from aiida.restapi.common.exceptions import RestValidationError, \
@@ -60,30 +61,6 @@ class CustomJSONEncoder(JSONEncoder):
 
         # If not returned yet, do it in the default way
         return JSONEncoder.default(self, o)
-
-
-class DatetimePrecision:
-    """
-    A simple class which stores a datetime object with its precision. No
-    internal check is done (cause itis not possible).
-
-    precision:  1 (only full date)
-                2 (date plus hour)
-                3 (date + hour + minute)
-                4 (dare + hour + minute +second)
-    """
-
-    def __init__(self, dtobj, precision):
-        """ Constructor to check valid datetime object and precision """
-
-        if not isinstance(dtobj, datetime):
-            raise TypeError('dtobj argument has to be a datetime object')
-
-        if not isinstance(precision, int):
-            raise TypeError('precision argument has to be an integer')
-
-        self.dtobj = dtobj
-        self.precision = precision
 
 
 class Utils:
